Break out coro.hpp
