+++ include/trompeloeil.hpp
@@ -33,5 +33,7 @@
 #include "trompeloeil/matcher/not.hpp"
 #include "trompeloeil/matcher/re.hpp"
 #include "trompeloeil/stream_tracer.hpp"
-
+#ifdef __cpp_impl_coroutine
+#include "trompeloeil/coro.hpp"
+#endif
 #endif // include guard
