+++ include/trompeloeil/mock.hpp
@@ -29,13 +29,6 @@
 // * Mocking function templates is not supported
 // * If a macro kills a kitten, this threatens extinction of all felines!
 
-#if defined(TROMPELOEIL_EXPERIMENTAL_COROUTINES)
-#  if defined(__cpp_impl_coroutine)
-#    define TROMPELOEIL_COROUTINES_SUPPORTED 1
-#  else
-#    error "Coroutines are not supported by this compiler"
-#  endif
-#endif
 
 #if defined(_MSC_VER)
 
@@ -145,6 +138,9 @@
 #include <type_traits>
 #include <utility>
 
+#ifdef __cpp_impl_coroutine
+#include <coroutine>
+#endif
 
 #ifndef TROMPELOEIL_CUSTOM_ATOMIC
 #include <atomic>
@@ -159,9 +155,6 @@ namespace trompeloeil { using std::unique_lock; }
 #include <trompeloeil/custom_unique_lock.hpp>
 #endif
 
-#ifdef TROMPELOEIL_COROUTINES_SUPPORTED
-#include <coroutine>
-#endif
 
 #ifdef TROMPELOEIL_SANITY_CHECKS
 #include <cassert>
@@ -495,45 +488,12 @@ namespace trompeloeil {
 
 # endif
 
-#ifdef TROMPELOEIL_COROUTINES_SUPPORTED
+#if defined(__cpp_impl_coroutine)
   template <typename Sig, typename = void>
   struct is_coroutine : std::false_type {};
 
   template <typename Sig>
   struct is_coroutine<Sig, std::void_t<decltype(&std::coroutine_traits<Sig>::promise_type::initial_suspend)>> : std::true_type {};
-
-  template <typename U>
-  struct type_wrapper{
-    template <typename V>
-    static V remove_rvalue_ref(V&&);
-    using type = decltype(remove_rvalue_ref(std::declval<U>()));
-  };
-  template <>
-  struct type_wrapper<void>
-  {
-    using type = void;
-  };
-
-  template <typename T>
-  struct coro_value_type
-  {
-    static auto func()
-    {
-      if constexpr (requires {std::declval<T>().operator co_await();})
-      {
-        return type_wrapper<decltype(std::declval<T>().operator co_await().await_resume())>{};
-      }
-      else
-      {
-        return type_wrapper<decltype(std::declval<T>().await_resume())>{};
-      }
-    }
-    using type = typename decltype(func())::type;
-  };
-
-template <typename T>
-using coro_value_type_t = typename coro_value_type<T>::type;
-
 #else
   template <typename>
   using is_coroutine = std::false_type;
@@ -1656,6 +1616,7 @@ template <typename T>
     return iterator{t};
   }
 
+
   class sequence_matcher;
 
   class sequence_type
@@ -2458,6 +2419,12 @@ template <typename T>
     }
   };
 
+  template <typename Sig, bool = is_coroutine<return_of_t<Sig>>::value>
+  struct yield_expr_base;
+
+  template <typename Sig>
+  using yield_expr_list = list<yield_expr_base<Sig>, delete_disposer>;
+
 #if TROMPELOEIL_MSVC
 #pragma warning(push)
 #pragma warning(disable : 4702)
@@ -2952,93 +2919,6 @@ template <typename T>
     T func;
   };
 
-#ifdef TROMPELOEIL_COROUTINES_SUPPORTED
-  template <typename Sig, bool = trompeloeil::is_coroutine<return_of_t<Sig>>::value>
-  struct yield_expr_base : public list_elem<yield_expr_base<Sig, true>>
-  {
-    ~yield_expr_base() override = default;
-
-    using ret_type = coro_value_type_t<return_of_t<Sig>>;
-    virtual
-    ret_type
-    expr(
-      call_params_type_t<Sig>&)
-    const = 0;
-  };
-  template <typename Sig>
-  struct yield_expr_base<Sig, false> : public list_elem<yield_expr_base<Sig, false>>
-  {
-    using ret_type = return_of_t<Sig>;
-    virtual
-    ret_type
-    expr(
-      call_params_type_t<Sig>&)
-    const = 0;
-  };
-
-  template <typename Sig>
-  using yield_expr_list = list<yield_expr_base<Sig>, delete_disposer>;
-
-  template <typename Sig, typename Expr>
-  struct yield_expr : yield_expr_base<Sig>
-  {
-    using ret_type = typename yield_expr_base<Sig>::ret_type;
-    template <typename E>
-    explicit
-    yield_expr(
-      E&& e_)
-      : e(std::forward<E>(e_))
-    {}
-
-    ret_type
-    expr(
-      call_params_type_t<Sig>& t)
-    const
-    override
-    {
-      return e(t);
-    }
-  private:
-    Expr e;
-  };
-
-  template <typename Sig, typename T>
-  class co_return_handler_t : public return_handler<Sig>
-  {
-  public:
-    template <typename U>
-    explicit
-    co_return_handler_t(
-      U&& u,
-      std::shared_ptr<yield_expr_list<Sig>> yields_)
-      : func(std::forward<U>(u))
-      , yields(std::move(yields_))
-    {}
-
-    return_of_t<Sig>
-    call(
-      trace_agent& /*agent*/,
-      call_params_type_t<Sig>& params)
-    override
-    {
-      using coro_type = return_of_t<Sig>;
-      using promise_type = typename std::coroutine_traits<coro_type>::promise_type;
-      using value_type = coro_value_type_t<coro_type>;
-      if constexpr (requires {std::declval<promise_type&>().yield_value(std::declval<value_type>());})
-      {
-        for (auto & e : *yields)
-        {
-          co_yield e.expr(params);
-        }
-      }
-      co_return func(params);
-    }
-  private:
-    T func;
-    std::shared_ptr<yield_expr_list<Sig>> yields;
-  };
-#endif
-
   template <typename Sig>
   class condition_base : public list_elem<condition_base<Sig>>
   {
@@ -3178,28 +3058,6 @@ template <typename T>
     static bool const sequence_set = true;
   };
 
-#ifdef TROMPELOEIL_COROUTINES_SUPPORTED
-  template <typename H, typename signature>
-    struct co_throw_handler_t
-    {
-      using R = decltype(default_return<return_of_t<signature>>());
-      using promise_value_type = trompeloeil::coro_value_type_t<R>;
-      explicit
-      co_throw_handler_t(H&& h_)
-        : h(std::move(h_))
-      {}
-
-      template <typename T>
-      promise_value_type operator()(T& p)
-      {
-        return ((void)h(p), trompeloeil::default_return<promise_value_type>());
-      }
-
-    private:
-      H h;
-    };
-#endif
-
   template <typename H, typename signature>
   struct throw_handler_t
   {
@@ -3333,92 +3191,6 @@ template <typename T>
     }
   };
 
-#ifdef TROMPELOEIL_COROUTINES_SUPPORTED
-  struct handle_co_yield
-  {
-    template <typename Matcher, typename modifier_tag, typename Parent, typename E>
-    static
-    call_modifier<Matcher, modifier_tag, Parent>
-    action(
-      call_modifier<Matcher, modifier_tag, Parent>&& m,
-      E&& e)
-    {
-      using signature = typename Parent::signature;
-      using params_type = call_params_type_t<signature>&;
-      using sigret = return_of_t<signature>;
-      using ret = std::invoke_result_t<E, params_type>;
-
-      constexpr bool is_void = std::is_same_v<ret, void>;
-      constexpr bool is_coroutine = trompeloeil::is_coroutine<sigret>::value;
-      constexpr bool is_matching_type = std::invoke([]{
-        if constexpr (is_coroutine && !is_void) {
-          using promise = typename std::coroutine_traits<sigret>::promise_type;
-          return requires(promise p, ret r){ p.yield_value(r); };
-        } else {
-          return false;
-        }
-      });
-      static_assert(is_coroutine,
-                    "CO_YIELD when return type is not a coroutine");
-      static_assert(!is_coroutine || !is_void,
-                    "You cannot CO_YIELD void");
-      static_assert(!is_coroutine || is_void || is_matching_type,
-                    "CO_YIELD is incompatible with the promise type");
-      constexpr auto valid = is_coroutine && !is_void && is_matching_type;
-      if constexpr (valid)
-      {
-        m.matcher->add_yield_expr(std::forward<E>(e));
-      }
-      return {m.matcher};
-    }
-  };
-  struct handle_co_return
-  {
-    template <typename Matcher, typename modifier_tag, typename Parent, typename H>
-    static
-    call_modifier<Matcher, modifier_tag, co_return_injector<return_of_t<typename Parent::signature>, Parent>>
-    action(call_modifier<Matcher, modifier_tag, Parent>&& m,
-           H&& h)
-    {
-      using signature = typename Parent::signature;
-      using return_type = typename Parent::return_type;
-      using co_return_type = typename Parent::co_return_type;
-      using params_type = call_params_type_t<signature>&;
-      using sigret = return_of_t<signature>;
-      using ret = std::invoke_result_t<H, params_type>;
-      constexpr bool has_return = !std::is_same_v<return_type, void>;
-      constexpr bool has_co_return = !std::is_same_v<co_return_type, void>;
-      constexpr bool is_coroutine = trompeloeil::is_coroutine<sigret>::value;
-      constexpr bool is_matching_type = std::invoke([]{if constexpr (is_coroutine) {
-        using promise = typename std::coroutine_traits<sigret>::promise_type;
-        if constexpr (std::is_same_v<void, ret>) {
-          return requires (promise p){p.return_void();};
-        } else {
-          return requires(promise p) { p.return_value(std::declval<ret>()); };
-        }}
-        return true;});
-
-      static_assert(!has_return,
-                    "CO_RETURN and RETURN cannot be combined");
-      static_assert(has_return || !has_co_return,
-                    "Multiple CO_RETURN does not make sense");
-      static_assert(has_return || is_coroutine, "CO_RETURN when return type is not a coroutine");
-      static_assert(has_return || has_co_return || !is_coroutine || is_matching_type,
-                    "Expression type does not match the coroutine promise type");
-      static_assert(!Parent::throws || Parent::upper_call_limit == 0,
-                    "CO_THROW and CO_RETURN does not make sense");
-      static_assert(Parent::upper_call_limit > 0,
-                    "CO_RETURN for forbidden call does not make sense");
-
-      constexpr bool valid = !has_return && !has_co_return && is_coroutine && is_matching_type && !Parent::throws && Parent::upper_call_limit > 0;
-      if constexpr (valid)
-      {
-        m.matcher->set_co_return(std::forward<H>(h));
-      }
-      return {m.matcher};
-    }
-  };
-#endif
   struct handle_return
   {
     template <typename Matcher, typename modifier_tag, typename Parent, typename H>
@@ -3487,42 +3259,6 @@ template <typename T>
       return {m.matcher};
     }
   };
-
-  #ifdef TROMPELOEIL_COROUTINES_SUPPORTED
-  struct handle_co_throw
-  {
-    template <typename Matcher, typename modifier_tag, typename Parent, typename H>
-    static
-    call_modifier<Matcher, modifier_tag, throw_injector<Parent>>
-    action(call_modifier<Matcher, modifier_tag, Parent>&& m,
-           H&& h)
-    {
-      using signature = typename Parent::signature;
-      using co_return_type = typename Parent::co_return_type;
-      using sigret = return_of_t<signature>;
-      constexpr bool is_coroutine      = trompeloeil::is_coroutine<sigret>::value;
-      static_assert(is_coroutine,
-                    "Do not use CO_THROW from a normal function, use THROW");
-      static_assert(!is_coroutine || !Parent::throws,
-                    "Multiple CO_THROW does not make sense");
-      constexpr bool has_return = !std::is_same<co_return_type, void>::value;
-      static_assert(!is_coroutine || !has_return,
-                    "CO_THROW and CO_RETURN does not make sense");
-
-      constexpr bool forbidden = Parent::upper_call_limit == 0U;
-
-      static_assert(!is_coroutine || !forbidden,
-                    "CO_THROW for forbidden call does not make sense");
-
-      constexpr bool valid = is_coroutine && !Parent::throws && !has_return;
-      if constexpr (valid) {
-        auto handler = co_throw_handler_t<H, signature>(std::forward<H>(h));
-        m.matcher->set_co_return(std::move(handler));
-      }
-      return {m.matcher};
-    }
-  };
-  #endif
 
   struct handle_throw
   {
@@ -3885,16 +3621,6 @@ template <typename T>
                                               std::forward<T>(t)...);
       sequences = std::move(seq);
     }
-#ifdef TROMPELOEIL_COROUTINES_SUPPORTED
-    template <typename E>
-    void
-    add_yield_expr(
-      E&& e)
-    {
-      auto expr = new yield_expr<Sig, E>(std::forward<E>(e));
-      yield_expressions->push_back(expr);
-    }
-#endif
     template <typename T>
     inline
     void
@@ -3913,24 +3639,10 @@ template <typename T>
     set_return(std::false_type, T&&t)//   RETURN
       noexcept;                      //   THROW
 
-#ifdef TROMPELOEIL_COROUTINES_SUPPORTED
-    template <typename T>
-    inline
-    void
-    set_co_return(
-      T&& h)
-    {
-      using basic_t = typename std::remove_reference<T>::type;
-      using handler = co_return_handler_t<Sig, basic_t>;
-      return_handler_obj.reset(new handler(std::forward<T>(h), yield_expressions));
-    }
-#endif
 
     condition_list<Sig>                    conditions;
     side_effect_list<Sig>                  actions;
-#ifdef TROMPELOEIL_COROUTINES_SUPPORTED
-    std::shared_ptr<yield_expr_list<Sig>>  yield_expressions = std::make_shared<yield_expr_list<Sig>>();
-#endif
+    std::shared_ptr<yield_expr_list<Sig>>  yield_expressions;
 
     std::unique_ptr<return_handler<Sig>>   return_handler_obj;
     std::unique_ptr<sequence_handler_base> sequences = detail::make_unique<sequence_handler<0>>();
@@ -4808,35 +4520,6 @@ template <typename T>
     return ::trompeloeil::decay_return_type(__VA_ARGS__);                      \
   })
 
-#ifdef TROMPELOEIL_COROUTINES_SUPPORTED
-
-#define TROMPELOEIL_CO_RETURN(...)    TROMPELOEIL_CO_RETURN_(=, __VA_ARGS__)
-#define TROMPELOEIL_LR_CO_RETURN(...) TROMPELOEIL_CO_RETURN_(&, __VA_ARGS__)
-
-#define TROMPELOEIL_CO_RETURN_(capture, ...)                                   \
-  template action<trompeloeil::handle_co_return>([capture](auto& trompeloeil_x)\
-                                                 -> decltype(auto) {           \
-    auto&& _1 = ::trompeloeil::mkarg<1>(trompeloeil_x);                        \
-    auto&& _2 = ::trompeloeil::mkarg<2>(trompeloeil_x);                        \
-    auto&& _3 = ::trompeloeil::mkarg<3>(trompeloeil_x);                        \
-    auto&& _4 = ::trompeloeil::mkarg<4>(trompeloeil_x);                        \
-    auto&& _5 = ::trompeloeil::mkarg<5>(trompeloeil_x);                        \
-    auto&& _6 = ::trompeloeil::mkarg<6>(trompeloeil_x);                        \
-    auto&& _7 = ::trompeloeil::mkarg<7>(trompeloeil_x);                        \
-    auto&& _8 = ::trompeloeil::mkarg<8>(trompeloeil_x);                        \
-    auto&& _9 = ::trompeloeil::mkarg<9>(trompeloeil_x);                        \
-    auto&&_10 = ::trompeloeil::mkarg<10>(trompeloeil_x);                       \
-    auto&&_11 = ::trompeloeil::mkarg<11>(trompeloeil_x);                       \
-    auto&&_12 = ::trompeloeil::mkarg<12>(trompeloeil_x);                       \
-    auto&&_13 = ::trompeloeil::mkarg<13>(trompeloeil_x);                       \
-    auto&&_14 = ::trompeloeil::mkarg<14>(trompeloeil_x);                       \
-    auto&&_15 = ::trompeloeil::mkarg<15>(trompeloeil_x);                       \
-    ::trompeloeil::ignore(_1,_2,_3,_4,_5,_6,_7,_8,_9,_10,_11,_12,_13,_14,_15); \
-    return ::trompeloeil::decay_return_type(__VA_ARGS__);                      \
-  })
-
-#endif
-
 #endif /* !(TROMPELOEIL_CPLUSPLUS != 201103L) */
 
 
@@ -4868,58 +4551,6 @@ template <typename T>
  })
 
 #endif /* !(TROMPELOEIL_CPLUSPLUS == 201103L) */
-
-#ifdef TROMPELOEIL_COROUTINES_SUPPORTED
-
-#define TROMPELOEIL_CO_THROW(...)    TROMPELOEIL_CO_THROW_(=, __VA_ARGS__)
-#define TROMPELOEIL_LR_CO_THROW(...) TROMPELOEIL_CO_THROW_(&, __VA_ARGS__)
-
-#define TROMPELOEIL_CO_THROW_(capture, ...)                                    \
-  template action<trompeloeil::handle_co_throw>([capture](auto& trompeloeil_x){\
-    auto&& _1 = ::trompeloeil::mkarg<1>(trompeloeil_x);                        \
-    auto&& _2 = ::trompeloeil::mkarg<2>(trompeloeil_x);                        \
-    auto&& _3 = ::trompeloeil::mkarg<3>(trompeloeil_x);                        \
-    auto&& _4 = ::trompeloeil::mkarg<4>(trompeloeil_x);                        \
-    auto&& _5 = ::trompeloeil::mkarg<5>(trompeloeil_x);                        \
-    auto&& _6 = ::trompeloeil::mkarg<6>(trompeloeil_x);                        \
-    auto&& _7 = ::trompeloeil::mkarg<7>(trompeloeil_x);                        \
-    auto&& _8 = ::trompeloeil::mkarg<8>(trompeloeil_x);                        \
-    auto&& _9 = ::trompeloeil::mkarg<9>(trompeloeil_x);                        \
-    auto&&_10 = ::trompeloeil::mkarg<10>(trompeloeil_x);                       \
-    auto&&_11 = ::trompeloeil::mkarg<11>(trompeloeil_x);                       \
-    auto&&_12 = ::trompeloeil::mkarg<12>(trompeloeil_x);                       \
-    auto&&_13 = ::trompeloeil::mkarg<13>(trompeloeil_x);                       \
-    auto&&_14 = ::trompeloeil::mkarg<14>(trompeloeil_x);                       \
-    auto&&_15 = ::trompeloeil::mkarg<15>(trompeloeil_x);                       \
-    ::trompeloeil::ignore(_1,_2,_3,_4,_5,_6,_7,_8,_9,_10,_11,_12,_13,_14,_15); \
-    throw __VA_ARGS__;                                                         \
- })
-
-#define TROMPELOEIL_CO_YIELD(...)    TROMPELOEIL_CO_YIELD_(=, __VA_ARGS__)
-#define TROMPELOEIL_LR_CO_YIELD(...) TROMPELOEIL_CO_YIELD_(&, __VA_ARGS__)
-
-
-#define TROMPELOEIL_CO_YIELD_(capture, ...)                                    \
-  template action<trompeloeil::handle_co_yield>([capture](auto& trompeloeil_x){\
-    auto&& _1 = ::trompeloeil::mkarg<1>(trompeloeil_x);                        \
-    auto&& _2 = ::trompeloeil::mkarg<2>(trompeloeil_x);                        \
-    auto&& _3 = ::trompeloeil::mkarg<3>(trompeloeil_x);                        \
-    auto&& _4 = ::trompeloeil::mkarg<4>(trompeloeil_x);                        \
-    auto&& _5 = ::trompeloeil::mkarg<5>(trompeloeil_x);                        \
-    auto&& _6 = ::trompeloeil::mkarg<6>(trompeloeil_x);                        \
-    auto&& _7 = ::trompeloeil::mkarg<7>(trompeloeil_x);                        \
-    auto&& _8 = ::trompeloeil::mkarg<8>(trompeloeil_x);                        \
-    auto&& _9 = ::trompeloeil::mkarg<9>(trompeloeil_x);                        \
-    auto&&_10 = ::trompeloeil::mkarg<10>(trompeloeil_x);                       \
-    auto&&_11 = ::trompeloeil::mkarg<11>(trompeloeil_x);                       \
-    auto&&_12 = ::trompeloeil::mkarg<12>(trompeloeil_x);                       \
-    auto&&_13 = ::trompeloeil::mkarg<13>(trompeloeil_x);                       \
-    auto&&_14 = ::trompeloeil::mkarg<14>(trompeloeil_x);                       \
-    auto&&_15 = ::trompeloeil::mkarg<15>(trompeloeil_x);                       \
-    ::trompeloeil::ignore(_1,_2,_3,_4,_5,_6,_7,_8,_9,_10,_11,_12,_13,_14,_15); \
-    return __VA_ARGS__;                                                         \
- })
-#endif
 
 
 #define TROMPELOEIL_TIMES(...) template action<trompeloeil::times>(::trompeloeil::multiplicity<__VA_ARGS__>{})
@@ -5029,14 +4660,6 @@ template <typename T>
 #define LR_RETURN                 TROMPELOEIL_LR_RETURN
 #define THROW                     TROMPELOEIL_THROW
 #define LR_THROW                  TROMPELOEIL_LR_THROW
-#ifdef TROMPELOEIL_COROUTINES_SUPPORTED
-#define CO_RETURN                 TROMPELOEIL_CO_RETURN
-#define LR_CO_RETURN              TROMPELOEIL_LR_CO_RETURN
-#define CO_THROW                  TROMPELOEIL_CO_THROW
-#define LR_CO_THROW               TROMPELOEIL_LR_CO_THROW
-#define CO_YIELD                  TROMPELOEIL_CO_YIELD
-#define LR_CO_YIELD               TROMPELOEIL_LR_CO_YIELD
-#endif
 #define TIMES                     TROMPELOEIL_TIMES
 #define IN_SEQUENCE               TROMPELOEIL_IN_SEQUENCE
 #define ANY                       TROMPELOEIL_ANY
